hiding agent status on a stopped workspaace
resolves #3484
