+++ site/src/components/WorkspaceActions/WorkspaceActions.test.tsx
@@ -1,8 +1,8 @@
 import { screen } from "@testing-library/react"
+import { WorkspaceStateEnum } from "util/workspace"
 import * as Mocks from "../../testHelpers/entities"
 import { render } from "../../testHelpers/renderHelpers"
 import { Language } from "./ActionCtas"
-import { WorkspaceStateEnum } from "./constants"
 import { WorkspaceActions, WorkspaceActionsProps } from "./WorkspaceActions"
 
 const renderComponent = async (props: Partial<WorkspaceActionsProps> = {}) => {
