+++ site/src/components/WorkspaceActions/WorkspaceActions.tsx
@@ -2,6 +2,7 @@ import Button from "@material-ui/core/Button"
 import Popover from "@material-ui/core/Popover"
 import { makeStyles } from "@material-ui/core/styles"
 import { FC, useEffect, useMemo, useRef, useState } from "react"
+import { WorkspaceStateEnum } from "util/workspace"
 import { Workspace } from "../../api/typesGenerated"
 import { getWorkspaceStatus, WorkspaceStatus } from "../../util/workspace"
 import { CloseDropdown, OpenDropdown } from "../DropdownArrows/DropdownArrows"
@@ -15,12 +16,7 @@ import {
   StopButton,
   UpdateButton,
 } from "./ActionCtas"
-import {
-  ButtonMapping,
-  ButtonTypesEnum,
-  WorkspaceStateActions,
-  WorkspaceStateEnum,
-} from "./constants"
+import { ButtonMapping, ButtonTypesEnum, WorkspaceStateActions } from "./constants"
 import { DropdownContent } from "./DropdownContent/DropdownContent"
 
 /**
